Remove pointless API bindings

+++ examples/llama3.py
@@ -4,14 +4,11 @@
 #
 
 from transformers import AutoTokenizer, TextStreamer
-from intel_npu_acceleration_library import NPUModelForCausalLM
-import torch
+from intel_npu_acceleration_library import NPUModelForCausalLM, int4
 
 model_id = "meta-llama/Meta-Llama-3-8B-Instruct"
 
-model = NPUModelForCausalLM.from_pretrained(
-    model_id, dtype=torch.int8, use_cache=True
-).eval()
+model = NPUModelForCausalLM.from_pretrained(model_id, dtype=int4, use_cache=True).eval()
 tokenizer = AutoTokenizer.from_pretrained(model_id)
 streamer = TextStreamer(tokenizer, skip_special_tokens=True, skip_prompt=True)
